Rename app to NewsOnTheGo and update banner logo
