+++ build.gradle
@@ -29,11 +29,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("seedu.newsonthego.NewsOnTheGo")
 }
 
 shadowJar {
-    archiveBaseName.set("duke")
+    archiveBaseName.set("newsonthego")
     archiveClassifier.set("")
 }
