+++ src/main/java/seedu/newsonthego/NewsOnTheGo.java
@@ -1,0 +1,23 @@
+package seedu.newsonthego;
+
+import java.util.Scanner;
+
+public class NewsOnTheGo {
+    /**
+     * Main entry-point for the java.duke.NewsOnTheGo application.
+     */
+    public static void main(String[] args) {
+        String logo = "#     #                         #######           #######                   ##### \n"
+                    + "##    # ###### #    #  ####     #     # #    #       #    #    # ######    #     #  ####  \n"
+                    + "# #   # #      #    # #         #     # ##   #       #    #    # #         #       #    # \n"
+                    + "#  #  # #####  #    #  ####     #     # # #  #       #    ###### #####     #  #### #    # \n"
+                    + "#   # # #      # ## #      #    #     # #  # #       #    #    # #         #     # #    # \n"
+                    + "#    ## #      ##  ## #    #    #     # #   ##       #    #    # #         #     # #    # \n"
+                    + "#     # ###### #    #  ####     ####### #    #       #    #    # ######     #####   #### \n";
+        System.out.println("Hello from\n" + logo);
+        System.out.println("What is your name?");
+
+        Scanner in = new Scanner(System.in);
+        System.out.println("Hello " + in.nextLine());
+    }
+}
