+++ src/test/java/seedu/newsonthego/NewsOnTheGoTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.newsonthego;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class NewsOnTheGoTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
